+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/XATransactionDataSourceWrapperTest.java
@@ -23,6 +23,7 @@ import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.constant.transaction.ProxyPoolType;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.transaction.xa.XATransactionManagerSPILoader;
+import io.shardingsphere.transaction.xa.manager.atomikos.XATransactionDataSourceWrapper;
 import io.shardingsphere.transaction.xa.manager.dialect.XADataSourceFactory;
 import io.shardingsphere.transaction.xa.manager.fixture.ReflectiveUtil;
 import io.shardingsphere.transaction.xa.manager.dialect.XADatabaseType;
